+++ src/runtime/adapters/ably/index.ts
@@ -1,5 +1,5 @@
 import type { SkewAdapterFactory } from '../types'
-import type { AblyAdapterConfig } from './types'
+import type { AblyAdapterConfig, AblyClientConfig } from './types'
 import { defineAdapter } from '../types'
 import { broadcast } from './node'
 import { ablyConfigSchema } from './types'
@@ -7,7 +7,7 @@ import { ablyConfigSchema } from './types'
 export type { AblyAdapterConfig } from './types'
 export { ablyConfigSchema } from './types'
 
-export const ablyAdapter: SkewAdapterFactory<AblyAdapterConfig> = defineAdapter({
+export const ablyAdapter: SkewAdapterFactory<AblyAdapterConfig, AblyClientConfig> = defineAdapter({
   name: 'ably',
   schema: ablyConfigSchema,
   clientModule: 'nuxt-skew-protection/adapters/ably/web',