+++ src/runtime/adapters/pusher/index.ts
@@ -1,5 +1,5 @@
 import type { SkewAdapterFactory } from '../types'
-import type { PusherAdapterConfig } from './types'
+import type { PusherAdapterConfig, PusherClientConfig } from './types'
 import { defineAdapter } from '../types'
 import { broadcast } from './node'
 import { pusherConfigSchema } from './types'
@@ -7,7 +7,7 @@ import { pusherConfigSchema } from './types'
 export type { PusherAdapterConfig } from './types'
 export { pusherConfigSchema } from './types'
 
-export const pusherAdapter: SkewAdapterFactory<PusherAdapterConfig> = defineAdapter({
+export const pusherAdapter: SkewAdapterFactory<PusherAdapterConfig, PusherClientConfig> = defineAdapter({
   name: 'pusher',
   schema: pusherConfigSchema,
   clientModule: 'nuxt-skew-protection/adapters/pusher/web',